Reverted the datetime begin & end code back to byte; reverted all dependent code. Changed the begin & end code to private. Modified databaseutil datetime methods to create a new object rather manipulating existing object. Fixed all the references to datetime public variables to respective methods.

+++ src/org/openelis/gwt/common/CalendarRPC.java
@@ -10,7 +10,7 @@ public class CalendarRPC implements RPC {
     public int month = -1;
     public int year = -1;
     public String monthDisplay;
-    public int begin;
-    public int end;
+    public byte begin;
+    public byte end;
 
 }
